Fix issue #299
After discovering, that PR #260 is reason for issue #299, i looked to its discussion, but could not find any mentions of exact problems, which appeared after menuitem functionality change. While it seems intended to make it more compact - imho it is not good idea to use negative valuds for margin.

p.s. Would be great of course to be able to reference to parrent of some widget class, not just child, as well as enhance selectors with specifier for arbitrary amount of elements in path, which satisfy to specified filter, implemented e.g. like list of selectors.